requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0

# Geospatial
rasterio>=1.3.6
shapely>=2.0.0
geopandas>=0.13.0
pyproj>=3.5.0
fiona>=1.9.0

# Azure
azure-storage-blob>=12.17.0
azure-identity>=1.13.0
azure-core>=1.27.0

# Scientific Computing
scipy>=1.10.0
scikit-image>=0.20.0
opencv-python>=4.7.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Monitoring
prometheus-client>=0.16.0
structlog>=23.1.0

# Performance
numba>=0.57.0
dask>=2023.5.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0